requirements.txt
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
black==20.8b1
blis==0.7.4
catalogue==1.0.0
certifi==2020.11.8
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
cymem==2.0.5
datefinder==0.7.1
Django==3.1.4
django-elasticsearch-dsl==7.1.4
django-elasticsearch-dsl-drf==0.20.9
django-filter==2.4.0
django-nine==0.2.4
djangorestframework==3.12.2
docopt==0.6.2
elasticsearch==7.9.1
elasticsearch-dsl==7.3.0
gunicorn==20.0.4
idna==2.10
joblib==1.0.0
lxml==4.6.1
Markdown==3.3.3
murmurhash==1.0.5
mypy-extensions==0.4.3
numpy==1.19.4
pathspec==0.8.1
plac==1.1.3
preshed==3.0.5
pyquery==1.4.1
python-dateutil==2.8.1
pytz==2020.5
regex==2020.11.13
requests==2.25.1
six==1.15.0
soupsieve==2.0.1
spacy==2.3.5
sqlparse==0.4.1
srsly==1.0.5
thinc==7.4.5
toml==0.10.2
tqdm==4.55.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
wasabi==0.8.0
Whoosh==2.7.4
